+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.0.5] - 2024-05-17
+
+### Added
+
+- PhoneNumber Provider
+- Added more examples to the `examples.py` file for testing
+
 ## [0.0.4] - 2024-05-16
 
 ### Added